LEMP Stack
This YAML file contains all the needed deployments for a LEMP deployment. Here we use Nginx to connect to Wordpress as a Frontend, and MySQL DB to be as a backend for the Stack. We have secrets being created here to use them as environment variables for the Deployments to connect with the MySQL setup, also to configure MySQL itself.